Former-commit-id: 3a0401cd2c4d33cc58e227c1eda620e7d1fdb544

+++ libraries/plumbing/backend_cuda/defs.h
@@ -136,7 +136,7 @@ using gpuError = hipError_t;
 
 namespace hila {
 __device__ __host__ double random();
-void seed_device_rng(unsigned long seed);
+void seed_device_rng(unsigned long long seed);
 } // namespace hila
 
 #define GPU_CHECK(cmd)                                                                 \
@@ -165,7 +165,7 @@ inline void synchronize_threads() {
 
 namespace hila {
 double random();
-void seed_device_rng(unsigned long seed);
+void seed_device_rng(unsigned long long seed);
 } // namespace hila
 
 using gpuError = int;
